Move `rotate_log` function to `talpid-core`
